src/gemini_integration.py
import google.generativeai as genai
import os

class GeminiIntegration:
    def __init__(self):
        api_key = os.getenv("GEMINI_API_KEY")
        if not api_key:
            raise ValueError("GEMINI_API_KEY environment variable not set.")
        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel("gemini-2.5-flash")

    def get_gemini_response(self, prompt: str) -> str:
        try:
            response = self.model.generate_content(prompt)
            return response.text
        except Exception as e:
            print(f"Could not call the Gemini API: {e}")
            return "Desculpe, ocorreu um erro ao me comunicar com a IA."