requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-docx
faiss-cpu
sentence-transformers
python-dotenv
numpy